Move credential provider adaptor to adaptor package
